Move cms components to models folder
