+++ projects/core/src/cms/store/effects/navigation-entry-item.effect.spec.ts
@@ -4,7 +4,6 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { cold, hot } from 'jasmine-marbles';
 import { Observable, of } from 'rxjs';
 
-import { CmsComponentList } from '../../../occ/occ-models/index';
 import { RoutingService } from '../../../routing/index';
 import * as fromEffects from './navigation-entry-item.effect';
 import * as fromActions from '../actions/navigation-entry-item.action';
@@ -12,7 +11,7 @@ import * as fromActions from '../actions/navigation-entry-item.action';
 import { provideMockActions } from '@ngrx/effects/testing';
 import { StoreModule } from '@ngrx/store';
 import * as fromCmsReducer from '../../../cms/store/reducers/index';
-import { OccConfig } from '@spartacus/core';
+import { CmsComponent, OccConfig } from '@spartacus/core';
 import { CmsComponentConnector } from '../../connectors/component/cms-component.connector';
 import { PageType } from '../../../model/cms.model';
 
@@ -48,7 +47,7 @@ class MockRoutingService {
 }
 
 class MockCmsComponentConnector {
-  getList(): Observable<CmsComponentList> {
+  getList(): Observable<CmsComponent[]> {
     return of(listComponents);
   }
 }
