+++ projects/core/src/occ/occ-models/index.ts
@@ -1,3 +1,2 @@
 export * from './additional-occ.models';
-export * from './cms-component.models';
 export * from './occ.models';
